Cleanup: Move math functions into utilities->math
